Refactor: Compare Page is now seperate component
